q: Ensafeify!
