+++ lib/probes/q.js
@@ -1,22 +1,17 @@
-var shimmer = require('shimmer')
-var tv = require('..')
+'use strict'
+
+const shimmer = require('shimmer')
+const tv = require('..')
+
+function patchThen (proto) {
+  if (typeof proto.then !== 'function') return
+  shimmer.wrap(proto, 'then', then => function (...args) {
+    return then.apply(this, args.map(tv.bind))
+  })
+}
 
 module.exports = function (q) {
-  shimmer.wrap(q.makePromise.prototype, 'then', function (then) {
-    return function (fulfilled, rejected, progressed) {
-      if (typeof fulfilled === 'function') {
-        fulfilled = tv.requestStore.bind(fulfilled)
-      }
-      if (typeof rejected === 'function') {
-        rejected = tv.requestStore.bind(rejected)
-      }
-      if (typeof progressed === 'function') {
-        progressed = tv.requestStore.bind(progressed)
-      }
-
-      return then.call(this, fulfilled, rejected, progressed)
-    }
-  })
-
+  const proto = q.makePromise && q.makePromise.prototype
+  if (proto) patchThen(proto)
   return q
 }
