+++ package.json
@@ -92,6 +92,7 @@
     "node-celery": "^0.2.3",
     "pg": "^4.4.1",
     "pg-native": "^1.10.0",
+    "q": "^1.4.1",
     "redis": "^2.5.3",
     "request": "^2.61.0",
     "restify": "^4.0.0",
