+++ test/probes/q.test.js
@@ -1,0 +1,6 @@
+var helper = require('../helper')
+var tv = helper.tv
+
+describe('probes/q', function () {
+  require('./promises')(require('q'))
+})
